[refactor] Filter가 전체 API에 적용 되던 오류 수정
